various changes

+++ src/app/games-management/pages/games-list/games-list.component.ts
@@ -4,7 +4,7 @@ import {State} from '../../../core/reducers';
 import {getGame} from '../../../game-management/store/current-game.selector';
 import {Observable, Subscription} from 'rxjs';
 import {Game} from '../../../game-management/store/current-game.state';
-import {DeleteGameRequestAction, GetGameListRequestAction, SetGamesFilterAction} from '../../store/game.actions';
+import {CloneGameRequestAction, DeleteGameRequestAction, GetGameListRequestAction, SetGamesFilterAction} from '../../store/game.actions';
 import {
     getFilteredGamesSelector,
     getGameList,
@@ -30,32 +30,32 @@ import {take} from "rxjs/operators";
             </div>
         </app-top-level-navbar>
         <div class="full-width-container maxwidth">
-            
-                <div class="gamesContainer-outer">
-                    <app-search-button
-                            placeholder="Start typing to search games ..."
-                            [dispatchAction]="dispatchAction"
-                            [filter]="filter"
 
-                    >
-                    </app-search-button>
-                    <div class="gamesContainer">
+            <div class="gamesContainer-outer">
+                <app-search-button
+                        placeholder="Start typing to search games ..."
+                        [dispatchAction]="dispatchAction"
+                        [filter]="filter"
 
-                        <app-screen-tile class="gameTile"
-                                         *ngFor="let game of (gameList$|async)"
-                                         [title]="game.title"
-                                         [subtitle]="game.lastModificationDate | date:'mediumDate'"
-                                         [imagePath]="game.splashScreen"
-                                         [actionText]="'VERWIJDER SPEL'"
-                                         [clickText]="'EDIT GAME'"
-                                         [navTo]="'/portal/game/' +game.gameId+ '/detail/screens'"
-                                         (actionClicked)="deleteGame(game)"
-                        ></app-screen-tile>
-                     
-                    </div>
+                >
+                </app-search-button>
+                <div class="gamesContainer">
+
+                    <app-screen-tile class="gameTile"
+                                     *ngFor="let game of (gameList$|async)"
+                                     [title]="game.title"
+                                     [subtitle]="game.lastModificationDate | date:'mediumDate'"
+                                     [imagePath]="game.splashScreen"
+                                     [actionText]="['GAME.DELETEGAME', 'GAME.CLONEGAME']"
+                                     [clickText]="'GAME.EDITGAME'"
+                                     [navTo]="'/portal/game/' +game.gameId+ '/detail/screens'"
+                                     (actionClicked)="deleteGame(game, $event)"
+                    ></app-screen-tile>
 
                 </div>
-            
+
+            </div>
+
         </div>
 
 
@@ -111,42 +111,6 @@ import {take} from "rxjs/operators";
             height: 388px;
             margin: 8px;
         }
-
-
-        /*.button-row {*/
-        /*    margin-bottom: 10px;*/
-        /*}*/
-        
-        /*.card-container {*/
-        /*    flex-wrap: nowrap;*/
-        /*    padding-left: 0;*/
-        /*    padding-right: 0;*/
-        /*    display: flex;*/
-        /*}*/
-        
-        /*.table-column {*/
-        /*    flex-grow: 3;*/
-        /*    flex-shrink: 1;*/
-        /*    width: 460px;*/
-        /*    margin: 10px;*/
-        
-        /*    min-width: 296px;*/
-        /*    flex-basis: 296px;*/
-        /*    !* background: gold; *!*/
-        /*}*/
-        
-        
-        /*.detail-column {*/
-        /*    flex-grow: 2;*/
-        /*    flex-shrink: 1;*/
-        /*    width: 460px;*/
-        /*    margin: 10px;*/
-        
-        /*    min-width: 296px;*/
-        /*    flex-basis: 296px;*/
-        /*    !* background: gold; *!*/
-        /*}*/
-
     `]
 })
 export class GamesListComponent implements OnInit, OnDestroy {
@@ -188,7 +152,11 @@ export class GamesListComponent implements OnInit, OnDestroy {
         }
     }
 
-    deleteGame(game) {
-        this.store.dispatch(new DeleteGameRequestAction({gameId: game.gameId}));
+    deleteGame(game, action) {
+        if (action === "GAME.DELETEGAME") {
+            this.store.dispatch(new DeleteGameRequestAction({gameId: game.gameId}));
+        } else if (action === "GAME.CLONEGAME") {
+            this.store.dispatch(new CloneGameRequestAction({gameId: game.gameId}));
+        }
     }
 }
